Integration: `2026.w02`

* !3828 L2 UE files code cleanup for static functions
* !3839 Couple rfsim fixes
* !3835 Correctly include cblas include directories for RHEL
* !3840 Fix missing return statements in SIB1 TDA validation
* !3836 move NTN UE L1 functions to a new file
* !3793 CI: Remove redundant Jenkinsfiles
* !3770 RRC UE: Implement proper NH chain synchronization for RRC Reestablishment
* !3843 fix(documentation): issue 1042
* !3842 Remove unused cmake_targets files
* !3832 Remove global iq_txshift and iq_rxrescale
* !3847 Minor CI and nr_ulsim reporting adjustments
* !3838 NR NGAP (N2HO) : Fix Time UE Stayed in Cell IE type and calculation
* !3677 Support prach with start symbol of PRACH not at symbol 0

Closes #1042 and #857

See merge request oai/openairinterface5g!3841

+++ openair2/COMMON/ngap_messages_types.h
@@ -473,8 +473,8 @@ typedef struct {
   cell_id_t id;
   // Cell type
   uint8_t type;
-  // Time UE Stayed in Cell
-  long time_in_cell;
+  // Time UE Stayed in Cell (seconds)
+  uint16_t time_in_cell;
   // Cause
   ngap_cause_t *cause;
 } last_visited_ngran_cell_info_t;
